+++ CHANGELOG.md
@@ -73,6 +73,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `jj workspace root` was aliased to `jj root`, for ease of discoverability
 
+* `jj diff` no longer shows the contents of binary files.
+
 ### Fixed bugs
 
 * Fixed snapshots of symlinks in `gitignore`-d directory.